refactor, add geojson
